requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
python-dotenv==0.15.0
pytz==2020.5
sqlparse==0.4.1
